pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "synciflow"
version = "0.0.1"
description = "Python package for the synciflow project."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "AdelElawady", email = "adel50ali50@gmail.com" },
]
keywords = []
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "typer>=0.12",
  "sqlmodel>=0.0.24",
  "pydantic>=2.0",
  "syncify-py>=1.0.0",
  "yt-dlp>=2025.1.0",
  "mutagen>=1.47.0",
  "ffmpeg-python>=0.2.0",
  "selenium>=4.0.0",
  "webdriver-manager>=4.0.0",
]

[project.scripts]
synciflow = "synciflow.cli.main:run"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.urls]
Homepage = "https://github.com/your-username/synciflow"
Repository = "https://github.com/your-username/synciflow"
Issues = "https://github.com/your-username/synciflow/issues"

[tool.hatch.version]
path = "src/synciflow/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
  "/README.md",
  "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/synciflow"]
